feat: restructure CLI application and dockerfile improvements
