+++ src/friendly_computing_machine/__init__.py
@@ -1,4 +1,4 @@
-from friendly_computing_machine.cli.cli import app
+from friendly_computing_machine.cli.main import app
 
 
 def main() -> None:
